adjust function position
